backend/requirements.txt
django
django-rest-framework
django-dotenv
mysqlclient
uuid
drf_yasg
django-cors-headers